PB - MF-277 - Move basic inputs to separate folders each
