+++ src/patient-registration/input/basic-input/checkbox/checkbox-input.component.tsx
