+++ src/patient-registration/input/basic-input/date/date-input.component.tsx
