+++ src/patient-registration/input/basic-input/number/number-input.component.tsx
