+++ src/patient-registration/patient-registration-validation.test.tsx
@@ -4,9 +4,9 @@ import { Formik, Form } from 'formik';
 import dayjs from 'dayjs';
 import { validationSchema } from './patient-registration-validation';
 import { NameInput } from './input/custom-input/name-input.component';
-import { SelectInput } from './input/basic-input/select-input.component';
-import { DateInput } from './input/basic-input/date-input.component';
-import { TelephoneNumberInput } from './input/basic-input/telephone-number-input/telephone-number-input.component';
+import { SelectInput } from './input/basic-input/select/select-input.component';
+import { DateInput } from './input/basic-input/date/date-input.component';
+import { TelephoneNumberInput } from './input/basic-input/telephone-number/telephone-number-input.component';
 import { EstimatedAgeInput } from './input/custom-input/estimated-age-input.component';
 
 describe('name input', () => {
